requirements.txt
asgiref==3.2.9
Django==3.0.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
